# contracts/platform/api/openapi.yaml
openapi: "3.1.0"
...
info:
  title: BlueLab v1 API
  version: 1.0.0
  summary: Voice-roleplay training and hiring platform — product, candidate, and operations surfaces.
  description: >
    Contract-first specification for the BlueLab application plane (F-1). Three principals:
    account holders (manager, rep) via session cookie; candidates via bearer capability token;
    BlueLab Internal Operations via a separate ops session. Cross-scope denial is an
    indistinguishable 404 (ADR-0036). All errors are RFC 9457 application/problem+json.
    Timestamps are UTC RFC 3339; months are YYYY-MM interpreted in the org's timezone;
    scores are 0-10 with one decimal and server-computed tier bands.
  license:
    name: Proprietary — BlueLab. All rights reserved.
    url: https://bluelab.example/license
  contact:
    name: BlueLab Engineering
servers:
  - url: https://api.bluelab.example
    description: Placeholder origin — Phase 7 owns real hosts. Paths carry their own /api/v1, /ops/v1, /hooks prefixes.
  - url: http://127.0.0.1:4010
    description: Prism mock server (contract walkthroughs).
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/authoring-options:
    get:
      tags: [Drills]
      summary: Challenge and hidden-motive libraries
      description: The provided selection libraries for authoring inputs (FR-DRL-002).
      operationId: listAuthoringOptions
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [challenge, hidden_motive] }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Active options.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/AuthoringOption" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills:
    post:
      tags: [Drills]
      summary: Create a draft drill
      description: >
        Creates a draft with the authoring intent (FR-DRL-002): call type (lead type iff Discovery,
        FR-DRL-001), one or more challenges and hidden motives (library or custom). Custom entries
        are job-relevance-checked at commit and rejected with the reason before they can influence
        generation (FR-DRL-007, AC-DRL-004, CMP-003). Language is fixed ar-EG (FR-DRL-003).
        Managers author team drills; reps author self-authored private drills (FR-TRP-009) — the
        mode follows the caller's role, no assignment step exists here.
      operationId: createDrill
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DrillInputs" }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "201":
          description: Draft created (FR-DRL-013 — resumable at any completeness).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DrillFull" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: "Problem types: validation-error; not-job-relevant (per-entry reasons in errors[])."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ValidationProblem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}:
    get:
      tags: [Drills]
      summary: Drill full basis (author / team manager)
      description: >
        The complete basis: inputs including the concealed set, generated scenario, rubric with
        weights, generation state, status. Reachable only by the drill's author and — for team
        drills — the owning team's manager (FR-TRM-010, FR-SCR-017 authorship lift; the author-only
        brief recap of FR-LIV-002 sources from here). Everyone else: 404 (ADR-0036). Published
        content is frozen — no mutation surface exists (FR-DRL-015, AC-DRL-006).
      operationId: getDrill
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Full drill.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DrillFull" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/inputs:
    put:
      tags: [Drills]
      summary: Replace authoring inputs (draft only)
      description: >
        Replaces the intent wholesale while the drill is a draft and invalidates its scenario,
        persona-derived label, rubric, captured grounding, and pending generation request
        (FR-DRL-002/005). Same job-relevance gate as creation (FR-DRL-007). On a published or
        archived drill: 409 drill-not-draft (FR-DRL-015). A displaced worker result is a no-op.
      operationId: replaceDrillInputs
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DrillInputs" }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Inputs replaced.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DrillFull" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem type: drill-not-draft."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: "Problem types: validation-error, not-job-relevant."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ValidationProblem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/scenario-generation:
    post:
      tags: [Drills]
      summary: Generate (or regenerate) the scenario
      description: >
        In one transaction, captures the team's current published facts and version vector, records
        a uniquely identified running request, invalidates the existing scenario, label, and rubric,
        and queues scenario generation (FR-DRL-004/005). The worker derives the unique buyer persona,
        situation, product references, and persona label from the stored inputs and captured grounding
        (FR-KNW-007), and writes only for the matching request. Requires at least one challenge and one
        motive (FR-DRL-002). Progress and failure surface on drill.generation (FR-DRL-006 — no fallback
        content; retry is a new request). 202 pattern: 00-contract-overview §8.
      operationId: generateScenario
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "202":
          description: Generation queued; poll GET /drills/{drill_id}.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerationAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem types: drill-not-draft, generation-in-progress."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: "Problem type: validation-error (missing challenge or motive)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ValidationProblem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/rubric-generation:
    post:
      tags: [Drills]
      summary: Generate (or regenerate) the rubric
      description: >
        Queues rubric generation (FR-DRL-008): dimensions with name, weight, rationale, derived from
        call type, challenges, motives, stored scenario, and the scenario request's captured grounding.
        When a rubric already exists the
        request must confirm that manual weight edits will be discarded (FR-DRL-011) — omitting
        discard_confirmed answers 422. An accepted request discards the existing rubric before work
        begins; only a complete validated result for the matching request installs the replacement.
      operationId: generateRubric
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                discard_confirmed:
                  type: boolean
                  description: Required true when regenerating over an existing rubric (FR-DRL-011).
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "202":
          description: Generation queued; poll GET /drills/{drill_id}.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerationAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem types: drill-not-draft, generation-in-progress, generation-incomplete (no scenario yet)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: "Problem type: validation-error (discard_confirmed required)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ValidationProblem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/rubric/weights:
    patch:
      tags: [Drills]
      summary: Tune rubric weights
      description: >
        Adjusts dimension weights while unpublished (FR-DRL-009a). Names and rationales are
        generation-fixed; no dimension can be added or reshaped by hand (AC-DRL-008). Response
        carries the live weight total for re-balancing display.
      operationId: tuneRubricWeights
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weights]
              properties:
                weights:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [dimension_id, weight]
                    properties:
                      dimension_id: { type: string, format: uuid }
                      weight: { type: integer, minimum: 0, maximum: 100 }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Updated rubric with live total.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RubricView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem type: drill-not-draft."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/rubric/dimensions/{dimension_id}:
    delete:
      tags: [Drills]
      summary: Delete a rubric dimension
      description: >
        Removes one dimension while unpublished (FR-DRL-009b, AC-DRL-008); publishing stays blocked
        until remaining weights re-balance to exactly 100 (FR-DRL-010).
      operationId: deleteRubricDimension
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
        - name: dimension_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Remaining rubric with live total.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RubricView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem type: drill-not-draft."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/publish:
    post:
      tags: [Drills]
      summary: Publish a drill (freeze)
      description: >
        The T-5 freeze transaction (FR-DRL-014): with successfully generated scenario and rubric
        states, a rubric totaling exactly 100, and a captured knowledge version vector that still
        exactly matches the team's live vector, promotes the captured grounding to the answer key
        and freezes scenario, persona-derived label, rubric, language, and answer key. The server
        computes the canonical content hash; all frozen content is immutable thereafter (FR-DRL-015).
        A knowledge-version mismatch answers 409 grounding-stale and requires scenario regeneration.
        Weights off 100 answer 409 weights-not-100 with meta.delta (AC-DRL-003 "+3 to balance").
      operationId: publishDrill
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Published drill (frozen).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DrillFull" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem types: weights-not-100 (meta.delta), generation-incomplete, grounding-stale, drill-not-draft."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/archive:
    post:
      tags: [Drills]
      summary: Archive a published drill
      description: >
        Withdraws the drill from all availability — libraries, assessments' future use, assignment
        (cancelling any outstanding assignment so it leaves recipients' libraries) — while every
        existing attempt, review, and statistic survives (FR-DRL-016, AC-DRL-007). Manager-only.
        Idempotent: archiving an archived drill answers 200 with current state.
      operationId: archiveDrill
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Archived drill.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DrillFull" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem type: drill-not-archivable (drafts cannot archive)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/brief:
    get:
      tags: [Drills]
      summary: Drill brief (participant-safe)
      description: >
        The pre-call brief rendered from the frozen scenario (FR-LIV-001): buyer identity block,
        situation context, product summary, content scaled to call type. For the drill's author the
        response additionally carries the collapsed setup recap (challenges + hidden motives,
        FR-LIV-002, AC-LIV-002) — the author_recap field is absent for every other viewer, and no
        other concealed element ever appears here (FR-SCR-017). Calls start only from this brief
        (FR-LIV-003). Published briefs are frozen content — cacheable (00-contract-overview §2).
      operationId: getDrillBrief
      security:
        - sessionCookie: []
        - candidateToken: []
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Brief.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BriefView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/drills/{drill_id}/reference:
    get:
      tags: [Drills]
      summary: Product reference for a drill
      description: >
        Read-only product reference reached from the brief (FR-KNW-009): the facts the drill was
        grounded in — the captured grounding for an author's generated draft and the frozen answer
        key for a published drill — grouped by document with the provenance line. What a participant
        studies equals what grading checks (FR-KNW-007). A draft without both a generated scenario and
        captured grounding answers 409 generation-incomplete. No edit affordance of any kind
        (AC-KNW-004).
      operationId: getDrillReference
      security:
        - sessionCookie: []
        - candidateToken: []
      parameters:
        - $ref: "#/components/parameters/DrillIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Reference.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferenceView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem type: generation-incomplete (draft scenario or grounding is unavailable)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # ───────────────────────────── Knowledge ─────────────────────────────
...
components:
  parameters:
...
    DrillIdParam:
      name: drill_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: "Problem type: session-invalid (missing/expired/revoked session)."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    OrgSuspended:
      description: "Problem type: organization-suspended. The confirmed service interval and current lifecycle override an otherwise valid ordinary grant."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: >
        Problem type: not-found. Also the uniform answer for any resource outside the caller's
        scope — indistinguishable from absence in status, body, and timing (ADR-0036,
        AC-IDA-006/007).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: "Problem type: validation-error, with field-level errors[]."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ValidationProblem" }
    RateLimited:
      description: "Problem type: rate-limited (05-rate-limits-and-quotas)."
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details; registry in 03-error-catalog.api.md (ADR-0036).
      required: [type, title, status]
      properties:
        type: { type: string, description: "Stable origin-relative problem URI, e.g. /problems/allowance-exhausted." }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        request_id: { type: string }
        meta:
          type: object
          description: Typed per-problem extension members (03-error-catalog per type).
          additionalProperties: true
      example:
        type: /problems/allowance-exhausted
        title: Attempts allowance exhausted
        status: 409
        detail: This assigned drill has used 3 of 3 allowed attempts.
        request_id: req_0197f3a1
        meta: { attempts_used: 3, attempts_allowed: 3 }
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
...
    CallType:
      type: string
      enum: [discovery, post_proposal, renewal, upsell]
      description: The four call types (FR-DRL-001).
    LeadType:
      type: string
      enum: [inbound_quote, referral, cold_outreach]
      description: Discovery-only lead type (FR-DRL-001).
...
    AuthoringOption:
      type: object
      required: [id, kind, label]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [challenge, hidden_motive] }
        label: { type: string }
    DrillInputs:
      type: object
      description: >
        Authoring intent (FR-DRL-002). Entries: source=library carries option_id; source=custom
        carries label (job-relevance-gated, FR-DRL-007). lead_type required iff call_type=discovery
        (FR-DRL-001). Generation requires ≥ 1 challenge and ≥ 1 hidden motive.
      required: [call_type, challenges, hidden_motives]
      properties:
        call_type: { $ref: "#/components/schemas/CallType" }
        lead_type:
          oneOf:
            - $ref: "#/components/schemas/LeadType"
            - type: "null"
        challenges:
          type: array
          items: { $ref: "#/components/schemas/ConcealedEntryInput" }
        hidden_motives:
          type: array
          items: { $ref: "#/components/schemas/ConcealedEntryInput" }
    ConcealedEntryInput:
      type: object
      required: [source]
      properties:
        source: { type: string, enum: [library, custom] }
        option_id: { type: [string, "null"], format: uuid, description: Required when source=library. }
        label: { type: [string, "null"], maxLength: 500, description: Required when source=custom. }
    ConcealedEntryView:
      type: object
      required: [source, label]
      properties:
        source: { type: string, enum: [library, custom] }
        label: { type: string }
    DrillFull:
      type: object
      description: >
        The full basis — author / owning-team manager only (FR-TRM-010, FR-SCR-017 authorship
        lift). Contains the concealed set; never serialized to any other viewer (the brief is the
        participant-safe projection).
      required: [id, status, self_authored, author_account_id, call_type, language, label, inputs, scenario, rubric, generation, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [draft, published, archived] }
        self_authored: { type: boolean, description: Rep-private drill (FR-TRP-009). }
        author_account_id: { type: string, format: uuid }
        call_type: { $ref: "#/components/schemas/CallType" }
        lead_type:
          oneOf:
            - $ref: "#/components/schemas/LeadType"
            - type: "null"
        language: { type: string, enum: [ar-EG], description: Fixed in v1 (FR-DRL-003). }
        label: { type: [string, "null"], description: Persona-derived; set by generation (FR-DRL-004). }
        inputs:
          type: object
          required: [challenges, hidden_motives]
          properties:
            challenges:
              type: array
              items: { $ref: "#/components/schemas/ConcealedEntryView" }
            hidden_motives:
              type: array
              items: { $ref: "#/components/schemas/ConcealedEntryView" }
        scenario:
          oneOf:
            - $ref: "#/components/schemas/ScenarioView"
            - type: "null"
        rubric:
          oneOf:
            - $ref: "#/components/schemas/RubricView"
            - type: "null"
        generation: { $ref: "#/components/schemas/GenerationState" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        published_at: { type: [string, "null"], format: date-time }
        archived_at: { type: [string, "null"], format: date-time }
...
    RubricView:
      type: object
      description: Dimensions with weights and live total; names/rationales generation-fixed (FR-DRL-008/009).
      required: [dimensions, total]
      properties:
        dimensions:
          type: array
          items:
            type: object
            required: [id, ord, name, weight, rationale]
            properties:
              id: { type: string, format: uuid }
              ord: { type: integer, minimum: 1 }
              name: { type: string }
              weight: { type: integer, minimum: 0, maximum: 100 }
              rationale: { type: string }
        total: { type: integer, description: Live weight total; publish requires exactly 100 (FR-DRL-010). }
...
    GenerationAccepted:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [queued] }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bluelab_session
      description: Account-holder server-side session (ADR-0037).
    candidateToken:
      type: http
      scheme: bearer
      description: Candidate capability token from the invite link fragment (FR-IDA-011, ADR-0037).
    opsSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bluelab_ops_session
      description: BlueLab Internal Operations session — separate surface (ADR-0010).
    livekitWebhook:
      type: apiKey
      in: header
      name: Authorization
      description: LiveKit-signed JWT carrying a SHA-256 of the payload (01-realtime-call-contract §7).